Begin split tmaze to cmaze, tmaze etc. & better errors
