+++ cmaze/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "cmaze"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+rand = "^0.8.5"
+rayon = "^1.5.3"
+thiserror = "1.0.31"
