+++ cmaze/src/core/mod.rs
@@ -1,0 +1,5 @@
+pub type Dims = (i32, i32);
+pub type Dims3D = (i32, i32, i32);
+#[allow(dead_code)]
+pub type DimsU = (usize, usize);
+pub type GameMode = (i32, i32, i32, bool);
