+++ cmaze/src/maze/mod.rs
@@ -1,0 +1,6 @@
+pub mod maze;
+pub use maze::Maze;
+pub mod cell;
+pub use cell::{Cell, CellWall};
+pub mod algorithms;
+pub use algorithms::*;
